--- app/src/main/java/com/exercise/matipv2/components/main/TotalTipAmount.kt
@@ -1,39 +1,0 @@
-package com.exercise.matipv2.components.main
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.dimensionResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import com.exercise.matipv2.R
-
-@Composable
-fun TotalTipAmount(finalTip: String) {
-    Row(
-        modifier = Modifier
-            .padding(bottom = dimensionResource(R.dimen.padding_sml))
-            .fillMaxSize(),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Text(
-            text = stringResource(R.string.tip_amount),
-            fontWeight = FontWeight.Bold,
-            style = MaterialTheme.typography.titleLarge,
-        )
-        Text(
-            modifier = Modifier.testTag("TipAmount"),
-            text = finalTip,
-            fontWeight = FontWeight.Bold,
-            style = MaterialTheme.typography.titleLarge,
-        )
-    }
-}
