+++ app/src/main/java/com/exercise/matipv2/ui/MainLayout.kt
@@ -27,8 +27,8 @@ import com.exercise.matipv2.R
 import com.exercise.matipv2.components.common.EditNumber
 import com.exercise.matipv2.components.common.RoundTheTipSwitch
 import com.exercise.matipv2.components.common.TopTabRow
-import com.exercise.matipv2.components.main.SplitCounter
-import com.exercise.matipv2.components.main.TotalTipAmount
+import com.exercise.matipv2.components.calculator.SplitCounter
+import com.exercise.matipv2.components.calculator.TotalTipAmount
 
 @SuppressLint("VisibleForTests")
 @Composable
